Predict method for fitted matching models

Calculate predicted values for matching models fitted with functions stabit and stabit2.

# S3 method for stabit2
predict(object, newdata = NULL, ...)

Arguments

object
a fitted object of class stabit

newdata
optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors or the fitted response values are returned.

...
.

References

Klein, T. (2015a). Does Anti-Diversification Pay? A One-Sided Matching Model of Microcredit. Cambridge Working Papers in Economics, #1521.

Examples

## load the results from Klein (2015) paper
 data(klein15a)

## predict the latent outcome variable
 predict(klein15a)

#>  [1] -3.6481798  3.0711275 -3.9270726  3.3712823  0.6507156  1.0714115
#>  [7] -2.4427181  1.4714352  0.1166499  0.7579598 -2.7648228  1.7544495
#> [13]  0.2947778 -0.7538749  0.8633761  2.2482714 -0.1241074  0.5437416
#> [19] -2.5142927  1.5967963 -2.1684489 -1.7951948 -3.2888203  2.5439853
#> [25] -2.3969827  1.3116390 -1.5986365 -1.8294208 -2.8371905  1.7916457
#> [31] -0.7752640 -2.1959717 -3.3915529  2.5571747 -2.7321937  1.8828920
#> [37] -1.5791602 -2.2540030  1.0478472 -2.0863312  1.3166756  2.3012064
#> [43] -2.2612972 -3.3525664 -2.9128948 -3.0470729 -3.2808990 -1.8860647
#> [49]  0.9059160  2.2928911 -3.4041963  2.5259717 -3.6928817  2.9508853
#> [55] -1.8582411 -3.9646937 -4.5822992 -2.6927001 -0.8212791 -7.9683328
#> [61] -2.7022713 -2.7342476  0.9611285  0.3111561 -0.5946951 -1.0904091
#> [67] -0.0922381 -2.2286941