data.frame for exploded logit model

shape a data.frame in a suitable form for the use of the exploded logit with the mlogit function in package mlogit.

xlogit.data(data, choice, shape = c("wide", "long"), varying = NULL,
  sep = ".", alt.var = NULL, chid.var = NULL, alt.levels = NULL,
  id.var = NULL, opposite = NULL, drop.index = FALSE,
  ranked = FALSE, ...)

Arguments

data
a data.frame

choice
the variable indicating the choice made: it can be either a logical vector, a numerical vector with 0 where the alternative is not chosen, a factor with level 'yes' when the alternative is chosen

shape
the shape of the data.frame: whether long if each row is an alternative or wide if each row is an observation

varying
the indexes of the variables that are alternative specific

sep
the seperator of the variable name and the alternative name (only relevant for a wide data.frame)

alt.var
the name of the variable that contains the alternative index (for a long data.frame only) or the name under which the alternative index will be stored (the default name is alt)

chid.var
the name of the variable that contains the choice index or the name under which the choice index will be stored

alt.levels
the name of the alternatives: if null, for a wide data.frame, they are guessed from the variable names and the choice variable (both should be the same), for a long data.frame, they are guessed from the alt.var argument

id.var
the name of the variable that contains the individual index if any

opposite
returns the opposite of the specified variables

drop.index
should the index variables be dropped from the data.frame

ranked
a logical value which is true if the response is a rank

...
further arguments passed to reshape

Value

xlogit.data returns a mlogit.data object, which is a data.frame in long format, i.e. one line for each alternative. It has a index attribute, which is a data.frame that contains the index of the choice made ('chid'), the index of the alternative ('alt') and, if any, the index of the individual ('id'). The choice variable is a boolean which indicates the choice made. This function use reshape if the data.frame is in wide format.

Examples

## Game2 is a data.frame in long format for which the response is a
## ranking variable (see example in package mlogit, function mlogit.data)

## --- 1. baseline case (equivalent to mlogit package)

## load data
data("Game2", package = "mlogit")
head(Game2, 13)

#>    age hours    platform ch own chid
#> 1   33  2.00     GameBoy  6   0    1
#> 2   33  2.00    GameCube  5   0    1
#> 3   33  2.00          PC  4   1    1
#> 4   33  2.00 PlayStation  1   1    1
#> 5   33  2.00  PSPortable  3   0    1
#> 6   33  2.00        Xbox  2   0    1
#> 7   19  3.25     GameBoy  6   0    2
#> 8   19  3.25    GameCube  5   0    2
#> 9   19  3.25          PC  1   1    2
#> 10  19  3.25 PlayStation  2   1    2
#> 11  19  3.25  PSPortable  3   0    2
#> 12  19  3.25        Xbox  4   0    2
#> 13  18  4.00     GameBoy  6   0    3

## transform data
G <- xlogit.data(Game2, shape = "long", choice = "ch",
                 alt.var = "platform", ranked = TRUE)
head(G, 30)

#>               age hours    platform own chid    ch
#> 1.GameBoy      33  2.00     GameBoy   0    1 FALSE
#> 1.GameCube     33  2.00    GameCube   0    1 FALSE
#> 1.PC           33  2.00          PC   1    1 FALSE
#> 1.PlayStation  33  2.00 PlayStation   1    1  TRUE
#> 1.PSPortable   33  2.00  PSPortable   0    1 FALSE
#> 1.Xbox         33  2.00        Xbox   0    1 FALSE
#> 2.GameBoy      33  2.00     GameBoy   0    1 FALSE
#> 2.GameCube     33  2.00    GameCube   0    1 FALSE
#> 2.PC           33  2.00          PC   1    1 FALSE
#> 2.PSPortable   33  2.00  PSPortable   0    1 FALSE
#> 2.Xbox         33  2.00        Xbox   0    1  TRUE
#> 3.GameBoy      33  2.00     GameBoy   0    1 FALSE
#> 3.GameCube     33  2.00    GameCube   0    1 FALSE
#> 3.PC           33  2.00          PC   1    1 FALSE
#> 3.PSPortable   33  2.00  PSPortable   0    1  TRUE
#> 4.GameBoy      33  2.00     GameBoy   0    1 FALSE
#> 4.GameCube     33  2.00    GameCube   0    1 FALSE
#> 4.PC           33  2.00          PC   1    1  TRUE
#> 5.GameBoy      33  2.00     GameBoy   0    1 FALSE
#> 5.GameCube     33  2.00    GameCube   0    1  TRUE
#> 6.GameBoy      19  3.25     GameBoy   0    2 FALSE
#> 6.GameCube     19  3.25    GameCube   0    2 FALSE
#> 6.PC           19  3.25          PC   1    2  TRUE
#> 6.PlayStation  19  3.25 PlayStation   1    2 FALSE
#> 6.PSPortable   19  3.25  PSPortable   0    2 FALSE
#> 6.Xbox         19  3.25        Xbox   0    2 FALSE
#> 7.GameBoy      19  3.25     GameBoy   0    2 FALSE
#> 7.GameCube     19  3.25    GameCube   0    2 FALSE
#> 7.PlayStation  19  3.25 PlayStation   1    2  TRUE
#> 7.PSPortable   19  3.25  PSPortable   0    2 FALSE

## --- 2. handle missing values 

## introduce NAs
Game2$ch[Game2$ch != 1] <- NA

## transform data
G <- xlogit.data(Game2, shape = "long", choice = "ch", alt.var = "platform", ranked = TRUE)
head(G, 30)

#>               age hours    platform own chid    ch
#> 1.GameBoy      33  2.00     GameBoy   0    1 FALSE
#> 1.GameCube     33  2.00    GameCube   0    1 FALSE
#> 1.PC           33  2.00          PC   1    1 FALSE
#> 1.PlayStation  33  2.00 PlayStation   1    1  TRUE
#> 1.PSPortable   33  2.00  PSPortable   0    1 FALSE
#> 1.Xbox         33  2.00        Xbox   0    1 FALSE
#> 2.GameBoy      19  3.25     GameBoy   0    2 FALSE
#> 2.GameCube     19  3.25    GameCube   0    2 FALSE
#> 2.PC           19  3.25          PC   1    2  TRUE
#> 2.PlayStation  19  3.25 PlayStation   1    2 FALSE
#> 2.PSPortable   19  3.25  PSPortable   0    2 FALSE
#> 2.Xbox         19  3.25        Xbox   0    2 FALSE
#> 3.GameBoy      18  4.00     GameBoy   0    3 FALSE
#> 3.GameCube     18  4.00    GameCube   0    3 FALSE
#> 3.PC           18  4.00          PC   1    3 FALSE
#> 3.PlayStation  18  4.00 PlayStation   1    3  TRUE
#> 3.PSPortable   18  4.00  PSPortable   0    3 FALSE
#> 3.Xbox         18  4.00        Xbox   0    3 FALSE
#> 4.GameBoy      18  6.00     GameBoy   1    4 FALSE
#> 4.GameCube     18  6.00    GameCube   1    4  TRUE
#> 4.PC           18  6.00          PC   1    4 FALSE
#> 4.PlayStation  18  6.00 PlayStation   0    4 FALSE
#> 4.PSPortable   18  6.00  PSPortable   0    4 FALSE
#> 4.Xbox         18  6.00        Xbox   1    4 FALSE
#> 5.GameBoy      18  2.00     GameBoy   0    5 FALSE
#> 5.GameCube     18  2.00    GameCube   0    5 FALSE
#> 5.PC           18  2.00          PC   1    5 FALSE
#> 5.PlayStation  18  2.00 PlayStation   0    5 FALSE
#> 5.PSPortable   18  2.00  PSPortable   0    5 FALSE
#> 5.Xbox         18  2.00        Xbox   0    5  TRUE

## --- 3. handle varying choice sets 

## drop choice options
Game2 <- Game2[-c(7:8),]

## transform data
G <- xlogit.data(Game2, shape = "long", choice = "ch", alt.var = "platform", ranked = TRUE)
head(G, 30)

#>               age hours    platform own chid    ch
#> 1.GameBoy      33  2.00     GameBoy   0    1 FALSE
#> 1.GameCube     33  2.00    GameCube   0    1 FALSE
#> 1.PC           33  2.00          PC   1    1 FALSE
#> 1.PlayStation  33  2.00 PlayStation   1    1  TRUE
#> 1.PSPortable   33  2.00  PSPortable   0    1 FALSE
#> 1.Xbox         33  2.00        Xbox   0    1 FALSE
#> 2.PC           19  3.25          PC   1    2  TRUE
#> 2.PlayStation  19  3.25 PlayStation   1    2 FALSE
#> 2.PSPortable   19  3.25  PSPortable   0    2 FALSE
#> 2.Xbox         19  3.25        Xbox   0    2 FALSE
#> 3.GameBoy      18  4.00     GameBoy   0    3 FALSE
#> 3.GameCube     18  4.00    GameCube   0    3 FALSE
#> 3.PC           18  4.00          PC   1    3 FALSE
#> 3.PlayStation  18  4.00 PlayStation   1    3  TRUE
#> 3.PSPortable   18  4.00  PSPortable   0    3 FALSE
#> 3.Xbox         18  4.00        Xbox   0    3 FALSE
#> 4.GameBoy      18  6.00     GameBoy   1    4 FALSE
#> 4.GameCube     18  6.00    GameCube   1    4  TRUE
#> 4.PC           18  6.00          PC   1    4 FALSE
#> 4.PlayStation  18  6.00 PlayStation   0    4 FALSE
#> 4.PSPortable   18  6.00  PSPortable   0    4 FALSE
#> 4.Xbox         18  6.00        Xbox   1    4 FALSE
#> 5.GameBoy      18  2.00     GameBoy   0    5 FALSE
#> 5.GameCube     18  2.00    GameCube   0    5 FALSE
#> 5.PC           18  2.00          PC   1    5 FALSE
#> 5.PlayStation  18  2.00 PlayStation   0    5 FALSE
#> 5.PSPortable   18  2.00  PSPortable   0    5 FALSE
#> 5.Xbox         18  2.00        Xbox   0    5  TRUE
#> 6.GameBoy      20 15.00     GameBoy   0    6 FALSE
#> 6.GameCube     20 15.00    GameCube   0    6 FALSE

# NOT RUN {
## --- 4. run models
summary(mlogit::mlogit(ch ~ own | -1 + hours + age, G, reflevel = "PC"))
summary(mlogit::mlogit(ch ~ 0 | -1 + age | own, G))
summary(mlogit::mlogit(ch ~ 0 | -1 + age, G))
# }